External PHP Update
