+++ README.md
@@ -9,40 +9,68 @@ This package has a much more niche use case than nvm does. When developing on Wi
 
 This utility changes that.
 
+---
+
 ## Installation
 
-Download the latest pvm version from the releases page (1.0-alpha-1, it's currently a pre-release).
+> [!WARNING]
+> version lower than 1.3.0 will have only pvm.exe
+> version 1.3.0 or higher will include pvm-setup.exe but can still get pvm.exe from source
 
-Create the folder `%UserProfile%\.pvm\bin` (e.g. `C:\Users\Harry\.pvm\bin`) and drop the pvm exe in there. Add the folder to your PATH.
+### Installer
+Download the latest pvm installer from the releases page (>= 1.3.0).
+
+### Manual Installation
+Create the folder `%UserProfile%\.pvm\bin` (e.g. `C:\Users\Harry\.pvm\bin`) and drop the pvm.exe in there. Add the folder to your PATH.
+
+---
 
 ## Commands
 ```
 pvm list
 ```
-Will list out all the available PHP versions you have installed
+ - __Will list out all the available PHP versions you have installed.__
 
 ```
-pvm path
+pvm list-remote
 ```
-Will tell you what to put in your Path variable.
+- __Will list available PHP versions from remote repositories.__
 
 ```
-pvm use 8.2.9
+pvm install <version> [nts] [path]
 ```
-> [!NOTE]  
-> Versions must have major.minor specified in the *use* command. If a .patch version is omitted, newest available patch version is chosen.
-
-Will switch your currently active PHP version to PHP 8.2.9
+- __Will install specified PHP vesrion.__
+- If the minor and patch versions are not specified, the newest available versions will be automatically selected.
+- [nts] (optional): Install a non-thread-safe version.
+- [path] (optional): Specify a custom installation path.
 
 ```
-pvm install 8.2
+pvm use <version|path>
 ```
-> [!NOTE]  
-> The install command will automatically determine the newest minor/patch versions if they are not specified
+- __Will switch your currently active PHP version to specified PHP vesrion.__
+-   Using a version: Specify at least the `major.minor` version. If the `.patch` version is omitted, the newest available patch version will be selected automatically.
+-   Using a path: Specify the path to the PHP executable to set it as the active PHP version.
 
-Will install PHP 8.2 at the latest patch.
+```
+pvm uninstall 8.2.9
+```
+- __Will uninstall specified PHP vesrion.__
+- The `uninstall` command requires the full `major.minor.patch` version to be specified.
+
+```
+pvm add <path>
+```
+- __Adds a custom PHP version by specifying the path to a PHP executable.__
+
+```
+pvm remove <path>
+```
+- __Removes a custom PHP version by specifying the path to the previously added PHP executable.__
+
+---
 
 ## Composer support
+
 `pvm` now installs also composer with each php version installed.
 It will install Composer latest stable release for PHP >= 7.2 and Composer latest 2.2.x LTS for PHP < 7.2.
 You'll be able to invoke composer from terminal as it is intended:
@@ -50,9 +78,18 @@ You'll be able to invoke composer from terminal as it is intended:
 composer --version
 ```
 
+---
+
 ## Build this project
 
 To compile this project use:
+
 ```shell
 GOOS=windows GOARCH=amd64 go build -o pvm.exe
+```
+
+To build pvm-setup.exe use:
+
+```shell
+iscc "pvm-setup.iss" 
 ```
