+++ src/go.mod
@@ -1,0 +1,17 @@
+module hjbdev/pvm
+
+go 1.18
+
+require (
+	github.com/fatih/color v1.13.0
+	github.com/stretchr/testify v1.9.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/mattn/go-colorable v0.1.9 // indirect
+	github.com/mattn/go-isatty v0.0.14 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/sys v0.0.0-20210630005230-0f9fa26af87c // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
